backend/lambda/python/requirements.txt
botocore==1.35.28
cryptography==43.0.1
mypy_boto3_dynamodb==1.35.24
mypy_boto3_secretsmanager==1.35.0
python-dotenv==1.0.1
validators==0.34.0
